keyframe function added
blender import trips test verios:
"set keyframe" function added
aninmation is note yet working
possibel problem: every circle that represents the bus at a specific
time and station seems to be a individual object. the keyframe is set,
but there is no animation since they are all different object.
also the keyframes are not yet linked to the actual time data but to a
dummy variable